Support for BMFF Hash V3 validation of fixedBlockSize and fixedBlockSizes Merkle trees generation of fixedBlockSize Merkle trees

+++ sdk/src/builder.rs
@@ -927,7 +927,7 @@ impl Builder {
         if dh.is_err() {
             let mut ph = DataHash::new("jumbf manifest", "sha256");
             for _ in 0..10 {
-                ph.add_exclusion(HashRange::new(0, 2));
+                ph.add_exclusion(HashRange::new(0u64, 2u64));
             }
             self.add_assertion(labels::DATA_HASH, &ph)?;
         }
@@ -1677,7 +1677,7 @@ mod tests {
 
         println!("offset: {}, size {}", offset, output_stream.get_ref().len());
         // create an hash exclusion for the manifest
-        let exclusion = crate::HashRange::new(offset, placeholder.len());
+        let exclusion = crate::HashRange::new(offset as u64, placeholder.len() as u64);
         let exclusions = vec![exclusion];
 
         let mut dh = DataHash::new("source_hash", "sha256");
